A......... [ZBX-5086] fixed updating templated host causes deleting of all non templated trigger-dependencies of this host
git-svn-id: svn://svn.zabbix.com/branches/dev@28051 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-5086/frontends/php/api/classes/CTrigger.php
@@ -1692,11 +1692,12 @@ class CTrigger extends CTriggerGeneral {
 	/**
 	 * Deletes all trigger dependencies from the given triggers and their children.
 	 *
-	 * @param array $triggers   an array of triggers with the 'triggerid' field defined
+	 * @param array $triggers	an array of triggers with the 'triggerid' field defined
+	 * @param bool $onlyTemplated	true to delete only dependencies from templates
 	 *
 	 * @return boolean
 	 */
-	public function deleteDependencies(array $triggers) {
+	public function deleteDependencies(array $triggers, $onlyTemplated = false) {
 		$triggers = zbx_toArray($triggers);
 
 		$this->validateDeleteDependencies($triggers);
@@ -1712,12 +1713,19 @@ class CTrigger extends CTriggerGeneral {
 				)
 			));
 			if ($childTriggers) {
-				$this->deleteDependencies($childTriggers);
+				$this->deleteDependencies($childTriggers, $onlyTemplated);
 			}
 
-			DB::delete('trigger_depends', array(
-				'triggerid_down' => $triggerids
-			));
+			if ($onlyTemplated) {
+				DBexecute("DELETE td FROM trigger_depends td
+					LEFT JOIN triggers t ON td.triggerid_up = t.triggerid
+					WHERE  templateid is not null AND".  DBcondition('triggerid_down', $triggerids));
+			}
+			else {
+				DB::delete('trigger_depends', array(
+					'triggerid_down' => $triggerids
+				));
+			}
 		}
 		catch (APIException $e) {
 			self::exception(ZBX_API_ERROR_PARAMETERS, _('Cannot delete dependency'));
@@ -1966,7 +1974,7 @@ class CTrigger extends CTriggerGeneral {
 						}
 					}
 				}
-				$this->deleteDependencies($childTriggers);
+				$this->deleteDependencies($childTriggers, true);
 
 				if ($newDependencies) {
 					$this->addDependencies($newDependencies);
